% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylEntropy.R
\name{methylEntropy}
\alias{methylEntropy}
\title{methylEntropy}
\usage{
methylEntropy(MF, N = length(MF))
}
\arguments{
\item{MF}{Methylation faction - a numeric vector}

\item{N}{Number of defaults to the length of MF}
}
\value{
entropy a numeric vector of length 1
}
\description{
Computes shannon entropy for methylation faction values
}
